refactor(context_compressor): extract durable session guards mixin (LB6)
Window 762-1203 (pin 1589-2030) into ContextCompressorDurableGuardsMixin.
Mixin composed BEFORE ContextEngine in bases so its on_session_end/start
shadow the base (as the godfile's own defs did) — MRO pitfall documented
in seam test. 15 members resolve is-identical through the class.
Deps: typing + time (stdlib). 70/70 targeted tests green.

Part of #78645
Part of #78647
