+++ tests/agent/test_summarize_tool_result_type_safety.py
@@ -127,7 +127,7 @@ class TestBackstopWrapper:
         """When a branch does fail, the fallback names the tool and size."""
         from unittest.mock import patch as _patch
         with _patch(
-            "agent.context_compressor._summarize_tool_result_unguarded",
+            "agent.context_compressor_summarizers._summarize_tool_result_unguarded",
             side_effect=TypeError("boom"),
         ):
             result = _summarize_tool_result("terminal", "{}", "y" * 300)
@@ -136,7 +136,7 @@ class TestBackstopWrapper:
     def test_backstop_handles_non_string_content(self):
         from unittest.mock import patch as _patch
         with _patch(
-            "agent.context_compressor._summarize_tool_result_unguarded",
+            "agent.context_compressor_summarizers._summarize_tool_result_unguarded",
             side_effect=TypeError("boom"),
         ):
             result = _summarize_tool_result("terminal", "{}", None)
